refactoring: moved ANSYS related stuff from MeshManager_Test to respective ANSYS model tests refactoring: rebuilding MeshManager_Test in a hierarchical fashion chore: removed BuildSurfaceConnectivty from AddFaces method because AGAIN it is the MeshManager's role to fix connectivity issues!

+++ source/includes/model/MeshManager.h
@@ -251,7 +251,7 @@ public:
                  typename std::vector<Face<dim>*>::iterator last );
 
   /// JCK's method to test the connectivity of a mesh after it had been read from binary file
-  int32_t CheckElementConnectivity() const;
+  size_t CheckElementConnectivity() const;
 
   /// prints stored objects and their connectivity to screen
   void Out() const;
